Update Test.js
Teachers figured out how to add a mark as submitted button themselves. But some forget to add it sometimes. You can use the post request from tasks with the button and it works fine for ones without the button if you change the ID. I may develop this so that it adds a mark as done button for tasks that wouldn't otherwise have one.